@@ -1,9 +1,5 @@
 declare global {
-	namespace App {}
-}
-
-declare module '@fortawesome/free-brands-svg-icon/index.es' {
-	export * from '@fortawesome/free-brands-svg-icons';
+  namespace App {}
 }
 
 export {};
